ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-ollama
langchain-text-splitters
sentence-transformers
faiss-cpu
pydf
python-multipart
python-dotenv
huggingface-hub
transformers
torch
accelerate
bitsandbytes
peft
trl
bytez